requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.5
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
flake8==3.9.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.2
progress==1.5
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.6
python-dotenv==0.16.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
